requirements.txt
better_exceptions==0.3.3
click==7.1.2
imageio==2.9.0
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.5
Pillow==8.4.0
plotly==5.1.0
requests==2.25.1
sacred==0.7.4
scikit_image==0.17.2
scikit_learn==0.24.2
scipy==1.5.2
setuptools==59.6.0
Shapely==1.7.1
streamlit==0.85.1
tensorflow-gpu==1.13.2
tqdm==4.55.0
opencv-python==4.0.1.23